% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_circle.R
\name{plotCircle}
\alias{plotCircle}
\title{Plot a circle}
\usage{
plotCircle(r = 1)
}
\arguments{
\item{r}{Radius of the circle}
}
\value{
None
}
\description{
Plot a circle
}
\examples{
plotCircle(2)
}

