% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/random.R
\name{randomNumber}
\alias{randomNumber}
\title{Generate random numbers}
\usage{
randomNumber(n, seed = 1234)
}
\arguments{
\item{n}{the number of random numbers to return}
}
\value{
a vector of random numbers of length n
}
\description{
Generate random numbers between 1 and 1000
}
\examples{
randomNumber(5)
}

